+++ src/main/java/com/github/jlangch/venice/impl/docgen/cheatsheet/modules/ModuleMimetypesSection.java
@@ -1,0 +1,56 @@
+/*   __    __         _
+ *   \ \  / /__ _ __ (_) ___ ___
+ *    \ \/ / _ \ '_ \| |/ __/ _ \
+ *     \  /  __/ | | | | (_|  __/
+ *      \/ \___|_| |_|_|\___\___|
+ *
+ *
+ * Copyright 2017-2024 Venice
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.jlangch.venice.impl.docgen.cheatsheet.modules;
+
+import com.github.jlangch.venice.impl.docgen.cheatsheet.DocItemBuilder;
+import com.github.jlangch.venice.impl.docgen.cheatsheet.DocSection;
+import com.github.jlangch.venice.impl.docgen.cheatsheet.ISectionBuilder;
+
+
+public class ModuleMimetypesSection implements ISectionBuilder {
+
+    public ModuleMimetypesSection(final DocItemBuilder diBuilder) {
+        this.diBuilder = diBuilder;
+    }
+
+    @Override
+    public DocSection section() {
+        final DocSection section = new DocSection(
+                                            "Mimetypes",
+                                            "modules.mimetypes");
+
+        final DocSection all = new DocSection("(load-module :mimetypes)", id());
+        section.addSection(all);
+
+        final DocSection hexdump = new DocSection("Mimetypes", id());
+        all.addSection(hexdump);
+        hexdump.addItem(diBuilder.getDocItem("mimetypes/probe-content-type", true));
+
+        return section;
+    }
+
+    private String id() {
+        return diBuilder.id();
+    }
+
+    private final DocItemBuilder diBuilder;
+}